Move unit tests to new directory
